just need to fix number of points in the homepage
