+++ src/components/Points/types.d.ts
@@ -1,0 +1,8 @@
+import { Profile } from '../../types/profile';
+
+export interface PointsProps {
+  isLoading: boolean;
+  isError: boolean;
+  error: Error | null;
+  data: Profile | undefined;
+}
